requirements/base.txt
aiohttp==3.4.4
gunicorn==19.7.1
simple-settings==0.12.0
uvloop==0.10.1
mongoengine
PyJWT